api-gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4